backend/pyproject.toml
[project]
name = "style-vault"
version = "0.1.0"
description = "Style Vault FastAPI backend"
readme = "README.md"
authors = [
    { name = "Links", email = "h96599021@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "fastapi[all]>=0.136.0",
    "loguru>=0.7.3",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
]

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ruff>=0.15.11",
]

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = ["*.pyc", "migrations"]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.isort]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
